First design for a decorator for easy versioning of functions using the schema_dict.
This allows the definition of functions that will be called instead of the default one based on the version numbers in the schema dictionary. It can also be used to define functions that are only available for specific versions if the default function raises an Error (e.g NotImplementedError).
The implementation is essentially singledispatch from functools but without types and with lambda functions instead :)

Docstrings can be a problem. When calling help on the function it will always show the default one. This will have to be kept in mind