Redesign of low level parsing functions to make them more userfriendly, when failure is encountered. Only low-level functions are changed in this commit. The file parsers are handled in the next one
The parser_info_out argument is removed from all low level functions and is replaced by either the functions always raising an error or a logger object which, when passed will take the errors and absorb the non-fatal errors for the parser. This should make it easier to reuse the schema_dict_util functions outside the file parsers, since they will raise an error immediately by default and we no longer need to know the structure the parser_info_out argument expects.