Add validation that the arguments given to modification registration methods on the FleurXMLModifier can actually be used to call the underlying XML modifying function. Here we use getcallargs from the inspect module for this.
Defining the signature in the modification would have been another valid approach. However, this gets very messy when extending modifying the functions and does not handle functions decorated with the schema_dict_version_dispatch cleanly. For example if the default function has some additional keyword arguments that are not valid for earlier versions we have to manually prevent these keyword arguments from being passed if the function is used for an older xml file