Simplify parts of write_inpgen_file
Formatting of positions/cell can be simplified more with f-strings, remove setting argument and pass the allowed options directly (The settings dictionary should be handled in aiida-fleur directly, otherwise a new allowed settings key would always need a new masci-tools release). path is renamed to file and is allowed to be a file handle that can be written to.