Make constants completely optional in schema_dict_util function
The reasoning is that 99% of users never define custom constants
and so for them the necessary reading in and passing around of the constants
from the xml-file is unecessary boilerplate. For the internal functions this is
completely fine but if you have to read in constants for everything you want to do
with no benefit for yourself this is annoying.

``calculate_expression`` now raises a custom exception for unknown
string expressions and uses the default constants defined for Fleur
if constants is not given.