Fix for compatibility with aiida-fleur in set_atomgroup_label
In version <=0.7.0 if an atom label did not exist the set_atomgroup_label
function passes silently. In 0.7.1 this started to raise an error as the new filters
version enforced the existence of the atom group. As for some internal use aiida-fleur
uses labels that may not exist we need to adjust the functions.