Allow referencing xml_getter functions in outxml parsing tasks
The parsing tasks predate the xml getters and at some points try to replicate behaviour,
which is very well defined and constructed in the xml_getters. In the next release of Fleur
there are some changes that would need to be taken into account to keep the film_relax_info
and bulk_relax_info tasks working, which is not worth it.

Instead we allow parse_type to be `xmlGetter` and give an argument name for the function