Make detection whether a xpath contains a tag more robust
previously the string was just checked for the containing tag
this hass a number of drawbacks including if the tag is in a Xpath
predicate or another part of the xpath contains the tag name to search

A function contains_tag is introduced to solve these problems. This
strips out the predicates and creates the explicit list of contained
tags.